Update getInt() function signature

+++ lib/get-num.c
@@ -60,7 +60,7 @@ long getLong(const char * arg, int flags, const char * name)
 }
 
 
-long getInt(const char * arg, int flags, const char * name)
+int getInt(const char * arg, int flags, const char * name)
 {
     long res;
